init healthBar script

+++ Assets/Scripts/HealthSystem.cs
@@ -17,7 +17,7 @@ public class HealthSystem
     }
 
     public float GetHealthPercent() {
-    	return health / healthMax;
+    	return (float)health / healthMax;
     }
 
     public void Damage(int damageAmount)
